pyproject.toml
[project]
name = "pdf-data-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.10",
    "langchain-openai>=0.2.11",
    "pypdf2>=3.0.1",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.0",
]
